+++ pkg/entity/grype-scan-report.go
@@ -1,0 +1,7 @@
+package entity
+
+import (
+	"github.com/anchore/grype/grype/presenter/models"
+)
+
+type GrypeScanReport models.Document