+++ pkg/exporter/defectDojo/models/paginated-response.go
@@ -1,0 +1,8 @@
+package models
+
+type PaginatedResponse struct {
+	Count    int           `json:"count"`
+	Next     string        `json:"next"`
+	Previous string        `json:"previous"`
+	Results  []interface{} `json:"results"`
+}