+++ pkg/exporter/exporter.go
@@ -1,0 +1,9 @@
+package exporter
+
+import (
+	"io"
+)
+
+type Exporter interface {
+	ExportGrype(reportFile io.Reader) error
+}